redirect users based on roles & flesh out admin + teacher pages
