+++ frontend/src/components/pages/TeacherPage.tsx
@@ -1,11 +1,20 @@
 import React from "react";
+
+import Logout from "../auth/Logout";
+import RefreshCredentials from "../auth/RefreshCredentials";
+import ResetPassword from "../auth/ResetPassword";
 import MainPageButton from "../common/MainPageButton";
 
 const TeacherPage = (): React.ReactElement => {
   return (
     <div style={{ textAlign: "center", width: "25%", margin: "0px auto" }}>
       <h1>Teacher Page</h1>
-      <MainPageButton />
+      <div className="btn-group">
+        <Logout />
+        <RefreshCredentials />
+        <ResetPassword />
+        <MainPageButton />
+      </div>
     </div>
   );
 };
